+++ Coimbra.Tests/EventSystemTests.cs
@@ -34,7 +34,7 @@ namespace Coimbra.Tests
         {
             const string log = nameof(log);
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });
@@ -49,12 +49,12 @@ namespace Coimbra.Tests
             const string logA = nameof(logA);
             const string logB = nameof(logB);
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(logA);
             });
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(logB);
             });
@@ -73,25 +73,25 @@ namespace Coimbra.Tests
             EventHandle handle3 = new();
             EventHandle handle4 = new();
 
-            void callback1(ref EventData<TestEvent> testEvent)
+            void callback1(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback1));
                 _eventService.RemoveListener(handle1);
             }
 
-            void callback2(ref EventData<TestEvent> testEvent)
+            void callback2(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback2));
                 _eventService.RemoveListener(handle2);
             }
 
-            void callback3(ref EventData<TestEvent> testEvent)
+            void callback3(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback3));
                 _eventService.RemoveListener(handle3);
             }
 
-            void callback4(ref EventData<TestEvent> testEvent)
+            void callback4(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(nameof(callback4));
                 _eventService.RemoveListener(handle4);
@@ -125,7 +125,7 @@ namespace Coimbra.Tests
         {
             const string log = nameof(log);
 
-            static void callback(ref EventData<TestEvent> testEvent)
+            static void callback(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(log);
             }
@@ -142,12 +142,12 @@ namespace Coimbra.Tests
             const string logA = nameof(logA);
             const string logB = nameof(logB);
 
-            static void callbackA(ref EventData<TestEvent> testEvent)
+            static void callbackA(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(logA);
             }
 
-            static void callbackB(ref EventData<TestEvent> testEvent)
+            static void callbackB(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(logB);
             }
@@ -175,12 +175,12 @@ namespace Coimbra.Tests
             const string logA = nameof(logA);
             const string logB = nameof(logB);
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(logA);
             });
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(logB);
             });
@@ -196,7 +196,7 @@ namespace Coimbra.Tests
             const string log = nameof(log);
             _eventService.SetEventKey<TestEvent>(new EventKey());
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });
@@ -213,7 +213,7 @@ namespace Coimbra.Tests
             EventKey eventKey = new();
             _eventService.SetEventKey<TestEvent>(eventKey);
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });
@@ -231,7 +231,7 @@ namespace Coimbra.Tests
         {
             const string log = nameof(log);
 
-            _eventService.AddListener(delegate(ref EventData<TestEvent> testEvent)
+            _eventService.AddListener(delegate(ref Event<TestEvent> testEvent)
             {
                 Debug.Log(log);
             });