+++ Roslyn~/Coimbra.Services.Events.Roslyn/EventDeclarationAnalyzer.cs
@@ -25,11 +25,11 @@ namespace Coimbra.Services.Events.Roslyn
 
             context.EnableConcurrentExecution();
             context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.None);
-            context.RegisterSyntaxNodeAction(x => AnalyzeObjectCreation<ClassDeclarationSyntax>(x, interfacePredicate), SyntaxKind.ClassDeclaration);
-            context.RegisterSyntaxNodeAction(x => AnalyzeObjectCreation<StructDeclarationSyntax>(x, interfacePredicate), SyntaxKind.StructDeclaration);
+            context.RegisterSyntaxNodeAction(x => AnalyzeEventDeclaration<ClassDeclarationSyntax>(x, interfacePredicate), SyntaxKind.ClassDeclaration);
+            context.RegisterSyntaxNodeAction(x => AnalyzeEventDeclaration<StructDeclarationSyntax>(x, interfacePredicate), SyntaxKind.StructDeclaration);
         }
 
-        private static void AnalyzeObjectCreation<T>(SyntaxNodeAnalysisContext context, Func<INamedTypeSymbol, bool> interfacePredicate)
+        private static void AnalyzeEventDeclaration<T>(SyntaxNodeAnalysisContext context, Func<INamedTypeSymbol, bool> interfacePredicate)
             where T : TypeDeclarationSyntax
         {
             if (context.Node is not T typeDeclarationSyntax
@@ -45,18 +45,18 @@ namespace Coimbra.Services.Events.Roslyn
             if (typeDeclarationSyntax.Parent is TypeDeclarationSyntax parentTypeNode)
             {
                 reportError = false;
-                context.ReportDiagnostic(Diagnostic.Create(Diagnostics.ConcreteEventShouldNotBeNested, typeDeclarationSyntax.GetLocation(), typeDeclarationSyntax.GetTypeName(), parentTypeNode.GetTypeName()));
+                context.ReportDiagnostic(Diagnostic.Create(Diagnostics.ConcreteEventShouldNotBeNested, typeDeclarationSyntax.Identifier.GetLocation(), typeDeclarationSyntax.GetTypeName(), parentTypeNode.GetTypeName()));
             }
 
             if (!typeDeclarationSyntax.Modifiers.Any(SyntaxKind.PartialKeyword))
             {
                 reportError = false;
-                context.ReportDiagnostic(Diagnostic.Create(Diagnostics.ConcreteEventShouldBePartial, typeDeclarationSyntax.GetLocation(), typeDeclarationSyntax.GetTypeName()));
+                context.ReportDiagnostic(Diagnostic.Create(Diagnostics.ConcreteEventShouldBePartial, typeDeclarationSyntax.Identifier.GetLocation(), typeDeclarationSyntax.GetTypeName()));
             }
 
             if (reportError && typeDeclarationSyntax.HasParameterlessConstructor(out bool isPublic) && !isPublic)
             {
-                context.ReportDiagnostic(Diagnostic.Create(Diagnostics.ConcreteEventParameterlessCtorShouldBePublic, typeDeclarationSyntax.GetLocation(), typeDeclarationSyntax.GetTypeName()));
+                context.ReportDiagnostic(Diagnostic.Create(Diagnostics.ConcreteEventParameterlessCtorShouldBePublic, typeDeclarationSyntax.Identifier.GetLocation(), typeDeclarationSyntax.GetTypeName()));
             }
         }
     }